refact. gunicorn 구동을 위한 서버 동작 리팩토링
