--- websocket/server/api.py
@@ -1,11 +1,0 @@
-from aiohttp import web
-
-
-routes = web.RouteTableDef()
-
-@routes.get('/')
-async def return_success(request):
-    return web.json_response({'rsp': 'success'})
-
-async def api_routes():
-    return routes
